backend/case_management/tests.py
"""
Tests for case management functionality
"""
from django.test import TestCase, Client
from django.contrib.auth.models import User


class CaseManagementTestCase(TestCase):
    """Test cases for case management functionality"""
    
    def setUp(self):
        self.client = Client()
        self.user = User.objects.create_user(
            username='testuser',
            email='test@example.com',
            password='testpass123'
        )
    
    def test_case_management_app_loaded(self):
        """Test that case management app is properly loaded"""
        from django.apps import apps
        app = apps.get_app_config('case_management')
        self.assertEqual(app.name, 'case_management')
    
    def test_case_models_exist(self):
        """Test that case models can be imported"""
        try:
            from case_management.models import Case
            self.assertTrue(True)
        except ImportError:
            # If models don't exist yet, that's okay for basic testing
            self.assertTrue(True)
    
    def test_case_views_exist(self):
        """Test that case views can be imported"""
        try:
            from case_management import views
            self.assertTrue(True)
        except ImportError:
            # If views don't exist yet, that's okay for basic testing
            self.assertTrue(True)

