backend/requirements-dev.txt
# Development and Testing Dependencies
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.5.3
factory-boy==3.3.0

# Code Quality
black==24.4.2
flake8==7.1.0
isort==5.13.2
mypy==1.10.0
django-stubs==5.0.2

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.4.2
ipython==8.25.0

# API Documentation
drf-spectacular==0.27.2

# Configuration
python-decouple==3.8

# Additional filters
django-filter==24.2

# Production server
gunicorn==22.0.0

